requirement.txt
Flask
python-dotenv
supabase
flask-cors
gunicorn
